Safari-609.1.11.1/LayoutTests/http/tests/adClickAttribution/anchor-tag-attributes-reflect.html

<!DOCTYPE html> <!-- webkit-test-runner [ internal:AdClickAttributionEnabled=true ] -->
<html lang="en">
<head>
    <meta charset="UTF-8">
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<a id="testTag" href="http://localhost:8000" adcampaignid="40" addestination="destination.example"></a>
<script>
    description("Test for the new ad click attribution attributes on anchor tags.");

    const anchorTag = document.getElementById("testTag");
    shouldBeEqualToString("anchorTag.adcampaignid", "40");
    shouldBeEqualToString("anchorTag.addestination", "destination.example");
</script>
</body>
</html>
