LayoutTests/http/tests/adClickAttribution/anchor-tag-attributes-reflect-expected.txt

Test for the new ad click attribution attributes on anchor tags.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS anchorTag.adcampaignid is "40"
PASS anchorTag.addestination is "destination.example"
PASS successfullyParsed is true

TEST COMPLETE
